+++ src/pages/Auth/SignIn.jsx
@@ -1,10 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import { useNavigate } from 'react-router-dom';
-
 import { login } from '../../api/Auth/user.js';
 import logo from '../../assets/svgs/logoex.svg';
-import KakaoLoginButton from '../../components/common/kakao-login/index.jsx';
+import KakaoButton from '../../components/common/kakao-button';
 import Input from '../../components/common/input/index.jsx';
 import FlexBox from '../../components/common/flex-box/index.jsx';
 
@@ -12,10 +11,6 @@ const SignIn = () => {
   const navigate = useNavigate();
 
   const [user, setUser] = useState({ email: '', password: '' });
-
-  const apiKey = import.meta.env.VITE_KAKAO_REST_API_KEY;
-  const redirectUri = import.meta.env.VITE_KAKAO_REDIRECT_URI;
-  const kakaoLoginPage = `https://kauth.kakao.com/oauth/authorize?response_type=code&client_id=${apiKey}&redirect_uri=${redirectUri}&response_type=code`;
 
   const handleLogin = async (e) => {
     e.preventDefault();
@@ -29,11 +24,6 @@ const SignIn = () => {
     } catch (error) {
       alert('이메일/비밀번호를 확인해주세요.');
     }
-  };
-
-  const handleKakaoLogin = (e) => {
-    e.preventDefault();
-    window.location.href = kakaoLoginPage;
   };
 
   const handleChange = (e) => {
@@ -72,9 +62,9 @@ const SignIn = () => {
         </FlexBox>
         <ButtonWrapper>
           <SignInButton onClick={handleLogin}>로그인</SignInButton>
-          <KakaoLoginButton full onClick={handleKakaoLogin}>
+          <KakaoButton full isLogin>
             카카오톡으로 계속하기
-          </KakaoLoginButton>
+          </KakaoButton>
           <NotAuth>아직 회원이 아니신가요?</NotAuth>
           <SignupButton onClick={() => navigate('/sign-up')}>
             회원가입
